BREAKING! Rename BaseVehicleController to VehiclePieceController to match it's purpose
