+++ src/ValheimVehicles/ValheimVehicles.Propulsion/Rudder/SteeringWheelComponent.cs
@@ -183,7 +183,7 @@ public class SteeringWheelComponent : MonoBehaviour, Hoverable, Interactable, ID
 
     if (user == Player.m_localPlayer)
     {
-      var baseVehicle = GetComponentInParent<BaseVehicleController>();
+      var baseVehicle = GetComponentInParent<VehiclePieceController>();
       if (baseVehicle != null)
       {
         baseVehicle.ComputeAllShipContainerItemWeight();
