+++ src/ValheimVehicles/ValheimVehicles.Vehicles/Controllers/WaterVehicleController.cs
@@ -19,7 +19,7 @@ using Logger = Jotunn.Logger;
 
 namespace ValheimVehicles.Vehicles;
 
-public class WaterVehicleController : BaseVehicleController, IWaterVehicleController
+public class WaterVehicleController : VehiclePieceController, IWaterVehicleController
 {
   private VehicleShip? _vehicleInstance;
